changing reading ini to double

+++ TankBoy/main.c
@@ -32,8 +32,8 @@ int main(void) {
     load_game_config(&game_system.config, "config.ini");
     
     // Create display with scaled buffer size
-    int disp_w = game_system.config.buffer_width * game_system.config.display_scale;
-    int disp_h = game_system.config.buffer_height * game_system.config.display_scale;
+    int disp_w = (int)(game_system.config.buffer_width * game_system.config.display_scale);
+    int disp_h = (int)(game_system.config.buffer_height * game_system.config.display_scale);
     ALLEGRO_DISPLAY* display = must_init(al_create_display(disp_w, disp_h), "display");
     ALLEGRO_EVENT_QUEUE* queue = must_init(al_create_event_queue(), "event queue");
